+++ LICENSE
@@ -1,0 +1,3 @@
+The use of this software is limited to education, research, and evaluation
+purposes only.  Commercial use is strictly prohibited.  For all other uses,
+contact the author(s).
