--- fast_lookup.h
@@ -1,9 +1,0 @@
-#ifndef __FAST_LOOKUP_H_
-#define __FAST_LOOKUP_H_
-#include "fast_lookup.h"
-
-
-void fast_table_init();
-
-extern BITMAP_TYPE fct[1<<STRIDE];
-#endif
