--- util.h
@@ -1,78 +1,0 @@
-#ifndef UTIL_H_
-#define UTIL_H_
-
-void out_of_memory();
-void *xcalloc(size_t count, size_t size);
-void *xzalloc(size_t size);
-void *xmalloc(size_t size);
-
-
-#ifdef UP_STATS
-extern int node_cpy;
-extern int node_alloc;
-extern int node_set;
-extern int st_f = 1;
-
-static inline void enable_stat()
-{
-    st_f = 1;
-}
-static inline void disable_stat() 
-{
-    st_f = 0;
-}
-#endif
-
-void mem_op_v6();
-
-/* Given a pointer-typed lvalue OBJECT, expands to a pointer type that may be
- * assigned to OBJECT. */
-#ifdef __GNUC__
-#define OVS_TYPEOF(OBJECT) typeof(OBJECT)
-#else
-#define OVS_TYPEOF(OBJECT) void *
-#endif
-
-/* Given OBJECT of type pointer-to-structure, expands to the offset of MEMBER
- * within an instance of the structure.
- *
- * The GCC-specific version avoids the technicality of undefined behavior if
- * OBJECT is null, invalid, or not yet initialized.  This makes some static
- * checkers (like Coverity) happier.  But the non-GCC version does not actually
- * dereference any pointer, so it would be surprising for it to cause any
- * problems in practice.
- */
-#ifdef __GNUC__
-#define OBJECT_OFFSETOF(OBJECT, MEMBER) offsetof(typeof(*(OBJECT)), MEMBER)
-#else
-#define OBJECT_OFFSETOF(OBJECT, MEMBER) \
-    ((char *) &(OBJECT)->MEMBER - (char *) (OBJECT))
-#endif
-
-/* Given POINTER, the address of the given MEMBER in a STRUCT object, returns
-   the STRUCT object. */
-#define CONTAINER_OF(POINTER, STRUCT, MEMBER)                           \
-        ((STRUCT *) (void *) ((char *) (POINTER) - offsetof (STRUCT, MEMBER)))
-
-/* Given POINTER, the address of the given MEMBER within an object of the type
- * that that OBJECT points to, returns OBJECT as an assignment-compatible
- * pointer type (either the correct pointer type or "void *").  OBJECT must be
- * an lvalue.
- *
- * This is the same as CONTAINER_OF except that it infers the structure type
- * from the type of '*OBJECT'. */
-#define OBJECT_CONTAINING(POINTER, OBJECT, MEMBER)                      \
-    ((OVS_TYPEOF(OBJECT)) (void *)                                      \
-     ((char *) (POINTER) - OBJECT_OFFSETOF(OBJECT, MEMBER)))
-
-/* Given POINTER, the address of the given MEMBER within an object of the type
- * that that OBJECT points to, assigns the address of the outer object to
- * OBJECT, which must be an lvalue.
- *
- * Evaluates to 1. */
-#define ASSIGN_CONTAINER(OBJECT, POINTER, MEMBER) \
-    ((OBJECT) = OBJECT_CONTAINING(POINTER, OBJECT, MEMBER), 1)
-
-
-
-#endif
